Remove docs duplication, update examples
