--- Examples/docs/MapsSearchView.swift
@@ -1,13 +1,0 @@
-import SwiftUI
-
-struct MapsSearchView: UIViewControllerRepresentable {
-    typealias UIViewControllerType = MapsSearchViewController
-    
-    func makeUIViewController(context: Context) -> MapsSearchViewController {
-        MapsSearchViewController()
-    }
-    
-    func updateUIViewController(_ uiViewController: MapsSearchViewController, context: Context) {
-        
-    }
-}
